Create Linked-List-Exercise-1.java
Problem Statement
Given a linked list, find and return the position of a given element in the linked list. If the element is not present, return 0. Implement the logic inside findPosition() method of the Tester class.

Note: The position starts from 1.

The LinkedList and Node classes have been provided.

 

Test the functionalities using the main() method of the Tester class.

Sample Input and Output