Create Stack-Assignment-1.java
Given a stack of integers, calculate the sum of all the integers present in the stack. Modify the stack such that the sum is present in the bottom of the stack and all the other integers are present in the stack in the same order. Implement the logic inside calculateSum() method of the Tester class.

The Stack class has been provided.

 

Test the functionalities using the main() method of the Tester class.